components: move Checkbox to form/
This matches what dashboard does and makes it easier to pull in things.
