+++ src/pages/K8s.vue
@@ -54,7 +54,7 @@
 </template>
 
 <script>
-import Checkbox from '@/components/Checkbox.vue';
+import Checkbox from '@/components/form/Checkbox.vue';
 import RadioGroup from '@/components/form/RadioGroup.vue';
 import SystemPreferences from '@/components/SystemPreferences.vue';
 const os = require('os');
